# domain-core-security-center-interfaces/src/main/java/com/firefly/security/center/interfaces/dtos/ContractInfoDTO.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.firefly.security.center.interfaces.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.UUID;

/**
 * Contract information including the party's role and associated product
 */
@Data
@Builder(toBuilder = true)
@AllArgsConstructor
@NoArgsConstructor
public class ContractInfoDTO {

    @JsonProperty("contractId")
    private UUID contractId;

    @JsonProperty("contractNumber")
    private String contractNumber;

    @JsonProperty("contractStatus")
    private String contractStatus;

    @JsonProperty("startDate")
    private LocalDateTime startDate;

    @JsonProperty("endDate")
    private LocalDateTime endDate;

    /**
     * Party's role in this contract
     */
    @JsonProperty("contractPartyId")
    private UUID contractPartyId;

    @JsonProperty("roleInContract")
    private RoleInfoDTO roleInContract;

    /**
     * Product linked to this contract
     */
    @JsonProperty("product")
    private ProductInfoDTO product;

    @JsonProperty("dateJoined")
    private LocalDateTime dateJoined;

    @JsonProperty("dateLeft")
    private LocalDateTime dateLeft;

    @JsonProperty("isActive")
    private Boolean isActive;

    @JsonProperty("createdAt")
    private LocalDateTime createdAt;

    @JsonProperty("updatedAt")
    private LocalDateTime updatedAt;
}