# domain-core-security-center-interfaces/src/main/java/com/firefly/security/center/interfaces/dtos/RoleInfoDTO.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.firefly.security.center.interfaces.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;
import java.util.UUID;

/**
 * Role information with associated scopes (permissions)
 */
@Data
@Builder(toBuilder = true)
@AllArgsConstructor
@NoArgsConstructor
public class RoleInfoDTO {

    @JsonProperty("roleId")
    private UUID roleId;

    @JsonProperty("roleCode")
    private String roleCode;

    @JsonProperty("name")
    private String name;

    @JsonProperty("description")
    private String description;

    @JsonProperty("isActive")
    private Boolean isActive;

    /**
     * List of role scopes (permissions) associated with this role
     */
    @JsonProperty("scopes")
    private List<RoleScopeInfoDTO> scopes;

    @JsonProperty("dateCreated")
    private LocalDateTime dateCreated;

    @JsonProperty("dateUpdated")
    private LocalDateTime dateUpdated;
}